\name{GetCoords}
\alias{GetCoords}
\title{Get coords in different coordinate systems}
\usage{
GetCoords(WKT, apiKey = .api, targetCoordSrid = 3021)
}
\arguments{
  \item{WKT}{WKT point from GetAddresses()}
}
\description{
Get coords in different coordinate systems from WKT (e.g.
RT90 2.5 from WGS 84)
}

