\name{GetAddresses}
\alias{GetAddresses}
\title{Get Street Adresses}
\usage{
  GetAddresses(apiKey = .api, municipalityPattern = "",
    streetName = "", streetNumPattern = "",
    postalCodePattern = "", postalAreaPattern = "",
    includeAddressConnectionsForTrafficTypes = "0")
}
\description{
  ...
}

